+++ packages/slate/src/index.ts
@@ -3,7 +3,6 @@ import { resetMemoization, useMemoization } from 'immutablejs-record-memoize';
 import Changes from './changes';
 import Block from './models/block';
 import Change from './models/change';
-import Character from './models/character';
 import Data from './models/data';
 import Document from './models/document';
 import History from './models/history';
@@ -33,17 +32,10 @@ Object.keys(Changes).forEach(type => {
     };
 });
 
-/*
- * Export.
- *
- * @type {Object}
- */
-
 export {
     Block,
     Change,
     Changes,
-    Character,
     Data,
     Document,
     History,
@@ -67,7 +59,6 @@ export {
 export default {
     Block,
     Changes,
-    Character,
     Data,
     Document,
     History,